+++ api/v1/cluster_types.go
@@ -641,7 +641,7 @@ const (
 	PhaseUnknownPlugin = "Cluster cannot proceed to reconciliation due to an unknown plugin being required"
 
 	// PhaseFailurePlugin is triggered when the cluster cannot proceed to reconciliation due to an interaction failure
-	PhaseFailurePlugin = "Cluster cannot proceed to reconciliation due to a failure while interacting with plugins"
+	PhaseFailurePlugin = "Cluster cannot proceed to reconciliation due to an error while interacting with plugins"
 
 	// PhaseImageCatalogError is triggered when the cluster cannot select the image to
 	// apply because of an invalid or incomplete catalog
@@ -668,6 +668,9 @@ const (
 
 	// PhaseCannotCreateClusterObjects is set by the operator when is unable to create cluster resources
 	PhaseCannotCreateClusterObjects = "Unable to create required cluster objects"
+
+	// PhaseNotEnoughSpace is set by the operator when there is not enough disk space and this is preventing postgres from starting
+	PhaseNotEnoughSpace = "Not enough disk space"
 )
 
 // EphemeralVolumesSizeLimitConfiguration contains the configuration of the ephemeral
